Code organization and reduce some breaking changes
