+++ example/lib/samples/modules/sample_one/navigation/sample_one_router.dart
@@ -13,7 +13,7 @@ const screenOneDeepLink = 'exapp://sampleOne/screenOne/id_1234_deepLink';
 @NuRouter()
 class SampleOneRouter extends Router {
   @override
-  String get prefix => 'exapp://sampleOne';
+  String get deepLinkPrefix => 'exapp://sampleOne';
 
   @NuRoute(deepLink: '/screenOne/:testId')
   ScreenRoute<String> screenOne({@required String testId}) => ScreenRoute(
