+++ lib/nuvigator.dart
@@ -1,7 +1,9 @@
 library nuvigator;
 
 export 'src/annotations.dart';
+export 'src/nu_route_settings.dart';
 export 'src/nuvigator.dart';
+export 'src/route_path.dart';
 export 'src/router.dart';
 export 'src/router_helpers.dart';
 export 'src/screen_route.dart';
