+++ lib/src/router_helpers.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 
+import 'route_path.dart';
 import 'router.dart';
 import 'screen_route.dart';
